OWRS-203 #time 1h 45m got full speed in both directions and stopped working for position controller in unit tests. Still need to test intermediate values and gears

+++ rover/src/owr_drive_controls/include/JointSpeedBasedPositionController.hpp
@@ -21,7 +21,7 @@ class JointSpeedBasedPositionController : public JointController {
         int maxRPM;
         
         int currentPWM;
-        double deltaPWM;
+        int deltaPWM;
         
         double maxVelocity;
         double velocityRange;
